move soundness test back
