+++ ceno_zkvm/src/e2e.rs
@@ -14,7 +14,6 @@ use crate::{
         ZKVMWitnesses,
     },
     tables::{MemFinalRecord, MemInitRecord, ProgramTableCircuit, ProgramTableConfig},
-    with_panic_hook,
 };
 use ceno_emul::{
     ByteAddr, CENO_PLATFORM, EmuContext, InsnKind, IterAddresses, Platform, Program, StepRecord,
@@ -24,14 +23,12 @@ use clap::ValueEnum;
 use ff_ext::{ExtensionField, GoldilocksExt2};
 use itertools::{Itertools, MinMaxResult, chain};
 use mpcs::{Basefold, BasefoldRSParams, PolynomialCommitmentScheme};
-use p3::{field::PrimeCharacteristicRing, goldilocks::Goldilocks};
+use p3::goldilocks::Goldilocks;
 use std::{
     collections::{BTreeSet, HashMap, HashSet},
-    panic,
-    panic::AssertUnwindSafe,
     sync::Arc,
 };
-use tracing::{error, info};
+use tracing::info;
 use transcript::{BasicTranscript as Transcript, BasicTranscriptWithStat, StatisticRecorder};
 
 pub type E = GoldilocksExt2;
@@ -672,10 +669,9 @@ fn format_segment(platform: &Platform, addr: u32) -> String {
 }
 
 pub fn verify(
-    zkvm_proof: ZKVMProof<E, Pcs>,
-    vk: ZKVMVerifyingKey<E, Pcs>,
+    zkvm_proof: &ZKVMProof<E, Pcs>,
+    verifier: &ZKVMVerifier<E, Pcs>,
 ) -> Result<(), ZKVMError> {
-    let verifier = ZKVMVerifier::new(vk);
     // print verification statistics like proof size and hash count
     let stat_recorder = StatisticRecorder::default();
     let transcript = BasicTranscriptWithStat::new(&stat_recorder, b"riscv");
@@ -685,44 +681,5 @@ pub fn verify(
         "hashes count = {}",
         stat_recorder.into_inner().field_appended_num
     );
-
-    // FIXME: it is a bit wired, let us move it else where later.
-    soundness_test(zkvm_proof, &verifier);
     Ok(())
-}
-
-fn soundness_test(mut zkvm_proof: ZKVMProof<E, Pcs>, verifier: &ZKVMVerifier<E, Pcs>) {
-    // do sanity check
-    let transcript = Transcript::new(b"riscv");
-    // change public input maliciously should cause verifier to reject proof
-    zkvm_proof.raw_pi[0] = vec![B::ONE];
-    zkvm_proof.raw_pi[1] = vec![B::ONE];
-
-    // capture panic message, if have
-    let result = with_panic_hook(Box::new(|_info| ()), || {
-        panic::catch_unwind(AssertUnwindSafe(|| {
-            verifier.verify_proof(zkvm_proof, transcript)
-        }))
-    });
-    match result {
-        Ok(res) => {
-            res.expect_err("verify proof should return with error");
-        }
-        Err(err) => {
-            let msg: String = if let Some(message) = err.downcast_ref::<&str>() {
-                message.to_string()
-            } else if let Some(message) = err.downcast_ref::<String>() {
-                message.to_string()
-            } else if let Some(message) = err.downcast_ref::<&String>() {
-                message.to_string()
-            } else {
-                unreachable!()
-            };
-
-            if !msg.starts_with("0th round's prover message is not consistent with the claim") {
-                error!("unknown panic {msg:?}");
-                panic::resume_unwind(err);
-            };
-        }
-    };
 }
